Simplify day 4 part 2 winner calculation
Previously `boards` was cloned in order to calculate the winners. Using
`Vec::retain` is much more readable for the same task, however it
requires us to allocate one more call to `is_board_solved` at the start
to handle logic if the last board is solved. This is more useful as it
improves code readability and is probably similar performance.